renaming quera_hardware to quera.
