+++ src/bloqade/builder/emit.py
@@ -331,7 +331,7 @@ class Emit(Builder):
     def __compile_hardware(
         self, nshots: int, backend: SubmissionBackend
     ) -> HardwareJob:
-        from bloqade.codegen.hardware.quera_hardware import SchemaCodeGen
+        from bloqade.codegen.hardware.quera import SchemaCodeGen
 
         capabilities = backend.get_capabilities()
 
@@ -386,7 +386,7 @@ class Emit(Builder):
         raise NotImplementedError
 
     def braket_local_simulator(self, nshots: int):
-        from bloqade.codegen.hardware.quera_hardware import SchemaCodeGen
+        from bloqade.codegen.hardware.quera import SchemaCodeGen
 
         if isinstance(self.register, ParallelRegister):
             raise TypeError("Braket emulator doesn't support parallel registers.")
